Add reproducible GSM8K-512 benchmark protocol and fix metric bugs
Adds a documented way to run this repo's guided-diffusion decoder as a
baseline and report accuracy, throughput, and tokens-per-step.

Fixes that change reported numbers:
- dataset.nshot was ignored; wrap_cot() always emitted all 8 CoT exemplars,
  so runs labelled 5-shot were actually 8-shot.
- tokens-per-step used (total_tokens + num_samples) / total_steps, but
  generated_length already counts the EOS token, inflating the metric.
- latency timing is now bracketed by torch.cuda.synchronize().

Additions:
- model.single_device pins both models to one GPU (the 2-GPU auto-split adds
  a cross-device copy to every verification step).
- model.dtype selects the load dtype (previously hardcoded float16).
- eval.warmup_samples runs untimed generations so CUDA init and kernel
  autotuning are not charged to sample 1.
- eval_results.json now records per-sample denoising_steps and ar_model_calls.
- scripts/summarize_flashdlm_metrics.py recomputes aggregate metrics.
- scripts/run_gsm8k_512_benchmark.sh runs the benchmark end to end.
- Two GSM8K-512 5-shot configs over the full 1319-problem test split.
- BENCHMARKING.md documents the protocol, metric definitions, and caveats.

+++ src/model/auto_map.py
@@ -103,7 +103,7 @@ class ModelMap:
         self.pretrained = pretrained
         self.peft = False
         
-    def fetch(self, device_map="auto"):
+    def fetch(self, device_map="auto", torch_dtype=torch.float16):
         print(f"Attempting to fetch model: '{self.model_name}'")
         print(f"Available models: {list(MODEL_LIBRARY_MAP.keys())}")
         if self.model_name not in MODEL_LIBRARY_MAP:
@@ -120,7 +120,7 @@ class ModelMap:
                             self.model_name,
                             config=config,
                             load_in_8bit=False,
-                            torch_dtype=torch.float16,
+                            torch_dtype=torch_dtype,
                             device_map=device_map,
                             trust_remote_code=True,
                         )
@@ -139,7 +139,7 @@ class ModelMap:
                             self.model_name,
                             config=config,
                             trust_remote_code=True,
-                            torch_dtype=torch.float16,
+                            torch_dtype=torch_dtype,
                         )
                     # Qwen2.5-Omni variants
                     # elif "Qwen2.5-Omni" in self.model_name:
@@ -148,14 +148,14 @@ class ModelMap:
                     #         self.model_name,
                     #         config=config,
                     #         trust_remote_code=True,
-                    #         torch_dtype=torch.float16,
+                    #         torch_dtype=torch_dtype,
                     #     )
                     #     return model
                     else:
                         model = AutoModelForCausalLM.from_pretrained(
                             self.model_name,
                             load_in_8bit=False,
-                            torch_dtype=torch.float16,
+                            torch_dtype=torch_dtype,
                             # device_map="auto",
                             device_map=device_map,
                             trust_remote_code=True,
